Update localization for React Native compatibility
