+++ package.json
@@ -25,6 +25,7 @@
     "expo-font": "~13.0.3",
     "expo-haptics": "~14.0.1",
     "expo-linking": "~7.0.5",
+    "expo-localization": "~16.0.1",
     "expo-router": "~4.0.17",
     "expo-splash-screen": "~0.29.21",
     "expo-status-bar": "~2.0.1",
@@ -33,14 +34,12 @@
     "expo-web-browser": "~14.0.2",
     "formik": "^2.4.5",
     "history": "^5.3.0",
-    "i18next": "^23.7.20",
-    "i18next-browser-languagedetector": "^7.2.0",
-    "i18next-http-backend": "^2.4.2",
+    "i18next": "^24.2.2",
     "lodash.debounce": "^4.0.8",
     "nativewind": "^4.1.23",
     "react": "18.3.1",
     "react-dom": "18.3.1",
-    "react-i18next": "^14.0.1",
+    "react-i18next": "^15.4.1",
     "react-native": "0.76.6",
     "react-native-drawer-layout": "^4.1.1",
     "react-native-gesture-handler": "~2.20.2",
