+++ src/components/RootContainer.web.tsx
@@ -1,10 +1,15 @@
-import React from 'react'
+import React, { useEffect } from 'react'
+import { i18n } from '@/i18n'
 import { View, StyleSheet } from 'react-native'
 export type Props = {
   children: React.ReactNode
 }
 
 const RootContainer: React.FC<Props> = ({ children }) => {
+  useEffect(() => {
+    document.dir = i18n.dir(i18n.language)
+  }, [])
+
   const styles = StyleSheet.create({
     rootView: {
       flex: 1,
